Added unique IDs and binary state
Unique IDs have been added. Although not necessary, it stops HA from complaining. 

Added an additional binary state if multi-mode arming is not enabled in the Inception, The armed state was not being detected (and the alarm could not be armed). I have made the payload to be ArmAway as it is essentially this mode and makes use of the existing buttons on the lovelace alarm panel.